+++ packages/component-library/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# v8.0.5 (July 24, 2026)
+
+### Added
+
+- Data Table component
+- Horizontal Stacked Bar component
+- FileDragUpload component
+
 # v8.0.4 (April 21, 2026)
 
 ### Removed